Refactor scrolling logic
